+++ finance/src/sgpp/finance/application/HestonSolver.cpp
@@ -84,7 +84,7 @@ HestonSolver::~HestonSolver() {
   }
 }
 
-void HestonSolver::constructGrid(base::BoundingBox& BoundingBox, int level) {
+void HestonSolver::constructGrid(base::BoundingBox& BoundingBox, size_t level) {
   this->dim = BoundingBox.getDimensions();
 
   if ((dim % 2) != 0)
