+++ pde/src/sgpp/pde/application/HeatEquationSolver.hpp
@@ -54,7 +54,7 @@ class HeatEquationSolver : public ParabolicPDESolver {
    */
   virtual ~HeatEquationSolver();
 
-  void constructGrid(sgpp::base::BoundingBox& myBoundingBox, int level);
+  void constructGrid(sgpp::base::BoundingBox& myBoundingBox, size_t level);
 
   virtual void solveExplicitEuler(size_t numTimesteps, double timestepsize, size_t maxCGIterations,
                                   double epsilonCG, sgpp::base::DataVector& alpha,
